Fix linter, refactor terrain
