+++ src/client/terrain/mod.rs
@@ -11,7 +11,7 @@ pub struct TerrainPlugin;
 impl Plugin for TerrainPlugin {
     fn build(&self, app: &mut App) {
         info!("Building TerrainPlugin");
-        app.insert_resource(terrain_resources::ChunkManager::new());
+        app.insert_resource(lib::ChunkManager::new());
         app.insert_resource(terrain_resources::SpawnAreaLoaded(false));
         app.insert_resource(util::TextureManager::new());
         app.add_event::<terrain_events::BlockUpdateEvent>();
